backend/requirements.txt
streamlit
langchain
langchain-community
langchain-pinecone
langchain-cohere
pinecone-client
cohere
python-dotenv
pypdf
